+++ src/manipulation_kit/executors/firmware/__init__.py
@@ -31,9 +31,10 @@ from .client import (FirmwareClient, hand_state, joint_state, lift_state,
                      neck_state)
 from .ensure import (ClientTree, Resolution, bundled_spec_sha256, ensure_client,
                      resolve_client, snapshot)
-from .errors import (ClientUnavailable, DeviceUnavailable, FirmwareError,
+from .errors import (ClientTimeout, ClientUnavailable, DeviceUnavailable, FirmwareError,
                      FirmwareUnavailable, LeasePreempted, ModeUnconfirmed,
-                     OperationUnavailable, ProtocolError, RateRefused, TrajectoryInvalid)
+                     OperationUnavailable, ProtocolError, RateRefused, RecoverFailed,
+                     TrajectoryInvalid)
 from .executor import (ANCHOR_GAP_DEG, DEFAULT_ACC_RATIO,
                        DEFAULT_STROKE_TIMEOUT_S, DEFAULT_TTL_S,
                        DEFAULT_VEL_RATIO, INTERPOLATION_S, LEASE_CLASS,
@@ -42,7 +43,7 @@ from .executor import (ANCHOR_GAP_DEG, DEFAULT_ACC_RATIO,
                        schedule_rate_deg_s)
 
 __all__ = [
-    "ANCHOR_GAP_DEG", "ClientTree", "ClientUnavailable",
+    "ANCHOR_GAP_DEG", "ClientTimeout", "ClientTree", "ClientUnavailable",
     "DEFAULT_ACC_RATIO", "DEFAULT_STROKE_TIMEOUT_S", "DEFAULT_TTL_S",
     "DEFAULT_VEL_RATIO",
     "DeviceUnavailable", "FirmwareClient", "FirmwareError", "FirmwareExecutor",
@@ -50,7 +51,7 @@ __all__ = [
     "Lease", "LeasePreempted", "MAX_COMMAND_STEP_DEG", "MAX_JOINT_RATE_DEG_S",
     "ModeUnconfirmed",
     "OperationUnavailable",
-    "ProtocolError", "RateRefused", "Resolution", "STREAM_HZ",
+    "ProtocolError", "RateRefused", "RecoverFailed", "Resolution", "STREAM_HZ",
     "bundled_spec_sha256", "default_holder", "ensure_client", "hand_state",
     "joint_state", "lift_state", "neck_state", "resolve_client",
     "TrajectoryInvalid", "schedule_rate_deg_s", "snapshot",